Added the UnityComponentProvider which allows you to map Unity Components + MonoBehaviours using the injector.
When an instance that had been mapped using the UnityComponentProvider is requested via injection, the dependency provider will attach the script to a GameObject using AddComponent instead of just trying to instantiate the class.
An example of a injector mapping using the UnityComponentProvider:
            context.injector.Map<IMyMonobehaviour>().ToProvider(new UnityComponentProvider(typeof(MyMonobehaviour)));

The UnityComponentProvider contains constructor overloads allowing you to configure the GameObject the script will be attached to.
If no GameObject is specified it will look for the ContextView GameObject.